Add a unit test for creating Articles
When an article is created it should download the contents from the
provided URI. This test verifies that behaviour. This is slightly
different from previous tests in that it is an 'integration test', using
an external system (a webpage). It might be worth adding an integration
test attribute/trait to this type of test later on.