ArticleSources should return a sequence of URIs, not Articles
This provides better decoupling of the source, and what we want to do with
the articles from it. As an ArticleSource conceptually represents the
start point of the pipeline, it should really just return a set of
pointers to the next places to visit. It also makes testing easier,
because at this point we don't have to worry about what happens if a web
request to one of the articles fails :)