Refactor common code out of the actor implementations
Lots of the logic of handling messages in actors is common (e.g.
cancellation, adding children, etc). Creating actors by passing in
functions representing the specific task for that actor reduces lots of
that duplication.

The current actor abstraction doesn't represent the differences between
'stateful' and 'stateless' actors very well, nor does it handle the
difference between Async and synchronous functions very well. This might
be something that needs to be revisited in the future.