Add word stemming algorithm
Stemming will allow recommendation algorithms to operate with greater
accuracy, as words which mean the same thing will be treated as the same,
rather than giving different rankings to words with the same root and the
same meaning. In the future, this could potentially be replaced with a lemmatizer for
more accuracy.