Remove punctuation and lower-case a string before stemming
The stemmer currently being used does not seem to remove punctuation. This
is a problem because it means words will not be stemmed to the same root
(or seemingly stemmed at all!) which will cause them to be processed
incorrect later on down the line.

A regression test has been added to illustrate this behaviour.