Create a Persistent Dictionary abstraction
The persistent dictionary will be one of a few 'Persistent Collections'
used to store the state of the application. The collections will be backed
by files, so the application can restore its state should it be shut down.

Conventionally this would be provided by a Database or other more formal
storage mechanism, but I feel like having simpler collections at this
point reduces complexity.