Create a PersistentSet implementation
This will be used to store the URIs which have already been processed,
meaning that a URI will never be processed twice, even if the application
shuts down in the meantime.

At some point it might be worth implementing an expiry mechanism, in case
URI content changes (or using storing something like a URI/Hash combo in
the set).