Add try/finally to PersistentCollections tests
This fixes a situation where a test failure means the backing file for a
persistent collection is not deleted. This will cause the test to fail
subsequently, as the file still exists and will therefore cause duplicate
key exceptions (or counts on queues to be incorrect, etc).

Moving the delete call to a finally block ensures its always cleaned up
appropriately. Ideally this would be moved to some form of disposal
method, but this doesn't seem to exist in FsUnit.