Create a 'State' object which contains all external application state
Simplify the persistence logic by creating one 'State' type which can be
initialized and used to carry out operations against storage.

This can be used for operations like getting the good/bad occurrences of a
word, which are needed for classification, without duplicating the logic
or implementing it at the top level.